+++ pallets/dynamic-energy/src/lib.rs
@@ -179,6 +179,10 @@ pub mod pallet {
     pub type EnergyGeneration<T: Config> =
         StorageMap<_, Twox64Concat, SessionIndex, EnergyOf<T>, ValueQuery>;
 
+    /// The minimum energy generation per session.
+    #[pallet::storage]
+    pub type MinimumEnergyGeneration<T: Config> = StorageValue<_, EnergyOf<T>>;
+
     /// The default smooth factor.
     /// Set to 1 to ignore past values and disable smoothing.
     #[pallet::type_value]
@@ -195,6 +199,8 @@ pub mod pallet {
         EnergySaleForceSet { amount: Option<EnergyOf<T>> },
         /// The total stake was forcibly set.
         TotalStakeForceSet { amount: Option<StakeOf<T>> },
+        /// The minimum energy generation was set.
+        MinimumEnergyGenerationSet { amount: Option<EnergyOf<T>> },
         /// The smooth factor for generation rate was updated.
         GenerationRateSmoothFactorUpdated { value: u32 },
         /// The smooth factor for exchange rate was updated.
@@ -340,6 +346,21 @@ pub mod pallet {
 
             Ok(())
         }
+
+        /// Set the minimum energy generation.
+        #[pallet::call_index(7)]
+        #[pallet::weight(T::DbWeight::get().writes(1))]
+        pub fn set_minimum_energy_generation(
+            origin: OriginFor<T>,
+            amount: Option<EnergyOf<T>>,
+        ) -> DispatchResult {
+            T::ManageOrigin::ensure_origin(origin)?;
+
+            MinimumEnergyGeneration::<T>::set(amount);
+            Self::deposit_event(Event::MinimumEnergyGenerationSet { amount });
+
+            Ok(())
+        }
     }
 }
 
@@ -456,7 +477,7 @@ impl<T: Config> Pallet<T> {
     }
 
     fn calculate_generation_rate(energy_burn: EnergyOf<T>) -> Option<EnergyOf<T>> {
-        Some(energy_burn)
+        Some(energy_burn.max(MinimumEnergyGeneration::<T>::get().unwrap_or_default()))
     }
 
     /// Calculates exchange rate by the following rule: